requirements.txt
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
joblib>=1.3.0
scipy>=1.11.0
